bpf: perf event header is 64bit
Therefore the rest of the message is also 64bit aligned.

+++ bpf/events/events.go
@@ -136,8 +136,8 @@ func (e *perfEventsReader) Map() bpf.Map {
 }
 
 type eventHdr struct {
-	Type uint16
-	Len  uint16
+	Type uint32
+	Len  uint32
 }
 
 type eventTimestampHdr struct {
